Remove Python library to move to the official library
