--- binding-library/python/azure/functions_extensions/kafka/__init__.py
@@ -1,7 +1,0 @@
-from .kafka import KafkaEvent, KafkaConverter, KafkaTriggerConverter
-
-__all__ = (
-    'KafkaEvent',
-    'KafkaConverter',
-    'KafkaTriggerConverter'
-)
