--- binding-library/python/azure/functions_extensions/kafka/_kafka.py
@@ -1,34 +1,0 @@
-import abc
-import typing
-
-
-class AbstractKafkaEvent(abc.ABC):
-
-    @abc.abstractmethod
-    def get_body(self) -> bytes:
-        pass
-
-    @property
-    @abc.abstractmethod
-    def key(self) -> typing.Optional[str]:
-        pass
-
-    @property
-    @abc.abstractmethod
-    def offset(self) -> typing.Optional[int]:
-        pass
-
-    @property
-    @abc.abstractmethod
-    def partition(self) -> typing.Optional[int]:
-        pass
-
-    @property
-    @abc.abstractmethod
-    def topic(self) -> typing.Optional[str]:
-        pass
-
-    @property
-    @abc.abstractmethod
-    def timestamp(self) -> typing.Optional[str]:
-        pass
