+++ samples/python/KafkaTrigger/main.py
@@ -1,5 +1,5 @@
 import logging
-from azure.functions_extensions.kafka import KafkaEvent
+from azure.functions import KafkaEvent
 
 def main(kevent : KafkaEvent):
     logging.info(kevent.get_body().decode('utf-8'))
