+++ samples/python/KafkaTriggerMany/main.py
@@ -1,6 +1,6 @@
 import logging
 import typing
-from azure.functions_extensions.kafka import KafkaEvent
+from azure.functions import KafkaEvent
 
 def main(kevents : typing.List[KafkaEvent]):
     for event in kevents:
