--- samples/python/library/azure_functions_kafka_binding.egg-info/PKG-INFO
@@ -1,48 +1,0 @@
-Metadata-Version: 2.1
-Name: azure-functions-kafka-binding
-Version: 1.0.1
-Summary: UNKNOWN
-Home-page: https://github.com/Azure/azure-functions-kafka-extension
-Author: Microsoft Corporation
-License: MIT License
-Description: # Azure Functions Python Binding for Kafka
-        This project provides Kafka binding for Azure Functions written in Python that makes it easy
-        to consume Kafka messages from one topic.
-        
-        # Features
-        - Reading from a Kafka topic
-        
-        # Getting Started
-        
-        ## Installation via PyPI
-        pip install azure-functions-kafka-binding
-        
-        ## Installation via Git
-        `git clone https://github.com/Azure/azure-functions-kafka-extension`
-        
-        `cd ./bindings_library/python/`
-        
-        `python setup.py install`
-        
-        ## Minimum Requirements
-        - Python 3.6
-        - See setup.py for dependencies
-        
-        # Installation
-        pip install azure-functions-kafka-binding
-        
-        # Usage
-        
-        ```python
-        from azure_functions.kafka import KafkaEvent
-        
-        def kafkaConsume(kevent: KafkaEvent):
-            print(kevent.get_body().decode('utf-8'))
-        ```
-        
-Platform: UNKNOWN
-Classifier: Development Status :: 3 - Alpha
-Classifier: Programming Language :: Python :: 3
-Classifier: Operating System :: OS Independent
-Classifier: License :: OSI Approved :: MIT License
-Description-Content-Type: text/markdown
