--- samples/python/library/azure_functions_kafka_binding.egg-info/dependency_links.txt
@@ -1,1 +1,0 @@
-
