--- samples/python/library/azure_functions_kafka_binding.egg-info/requires.txt
@@ -1,1 +1,0 @@
-azure-functions>=1.0.5
