--- samples/python/library/azure_functions_kafka_binding.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-azure
