+++ samples/python/requirements.txt
@@ -1,0 +1,1 @@
+azure.functions==1.3.1
